backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Image processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# SAM 3 and ML
torch==2.2.0
torchvision==0.17.0
# Note: Install SAM3 separately from https://github.com/facebookresearch/sam2
# pip install git+https://github.com/facebookresearch/sam2.git

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
cachetools==5.3.2

# Development
pytest==7.4.4
httpx==0.26.0
